+++ CHANGELOG.md
@@ -18,6 +18,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     - Step next (Ctrl+DownArrow)
         - Key/value -> next key-value pair in object
         - Array item -> next array item
-    - Step previous (Ctrl+DownArrow)
+    - Step previous (Ctrl+UpArrow)
         - Key/value -> previous key-value pair in object
         - Array item -> previous array item
+
+## 0.2.0
+
+- 2 new Commands (Keyboard shortcut)
+    - Include next (Ctrl+Shift+DownArrow)
+        - Key/value -> include the next key-value pair in the selection
+        - Array item -> include the next array item in the selection
+    - Include previous (Ctrl+Shift+UpArrow)
+        - Key/value -> include the previous key-value in the selection
+        - Array item -> include the previous array item in the selection
+- Bugfixes
+    - When using "Step next" at the last property of an object both key and value is not selected as expected